fix: calculator example/test return type
There's a major discrepancy with how `2025-06-18` handles
`structuredContent` in relation to `outputSchema` being defined as JSON
Schema.

Because `structuredContent` must be an `object`, the types that can
be returned by servers and validated by clients (like `type: array`
which is fully supported by JSON Schema but throws errors in Inspector,
the Typescript-SDK, etc) is dramatically narrow.

Related MCP issue:

https://github.com/modelcontextprotocol/modelcontextprotocol/issues/834

To deal with this, the calculator example/test tools now use a struct
return type.